+++ api/src/main/java/org/openmrs/module/poc/api/common/util/OPENMRSUUIDs.java
@@ -46,4 +46,6 @@ public interface OPENMRSUUIDs {
 	String RAPID_PLASMA_REAGIN_CONCEP_UUID = "e1dd332c-1d5f-11e0-b929-000c29ad1d07";
 	
 	String DVRL_CONCEPT_UUID = "e1d14ce2-1d5f-11e0-b929-000c29ad1d07";
+	
+	String LABORATORIO_GERAL_META_FORM_UUID = "e28b23a6-1d5f-11e0-b929-000c29ad1d07";
 }
