+++ omod/src/main/java/org/openmrs/module/poc/web/resource/TestOrderResultResource.java
@@ -18,8 +18,8 @@ import java.util.List;
 import org.apache.commons.lang3.StringUtils;
 import org.openmrs.Encounter;
 import org.openmrs.api.context.Context;
-import org.openmrs.module.poc.testresult.model.TestOrderResult;
-import org.openmrs.module.poc.testresult.service.TestOrderResultService;
+import org.openmrs.module.poc.testorderresult.model.TestOrderResult;
+import org.openmrs.module.poc.testorderresult.service.TestOrderResultService;
 import org.openmrs.module.webservices.rest.web.RequestContext;
 import org.openmrs.module.webservices.rest.web.RestConstants;
 import org.openmrs.module.webservices.rest.web.annotation.PropertyGetter;
